requirements.txt
black>=23.9.1
pre-commit>=3.5.0
ruff>=0.0.292
yt-dlp
ffmpeg-python
whispercpp
faster-whisper